Add protobufs with Wire
