--- build.gradle
@@ -1,36 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.jetbrains.kotlin.jvm' version '1.4.10'
-}
-
-group 'com.caseykulm'
-version '0.1'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-    jcenter()
-}
-
-dependencies {
-    implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
-    implementation("com.squareup.moshi:moshi:1.11.0")
-    implementation("com.squareup.moshi:moshi-kotlin-codegen:1.11.0")
-    implementation("com.squareup.okio:okio:2.9.0")
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.9")
-
-    testImplementation("org.junit.jupiter:junit-jupiter-api:5.1.0")
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.1.0")
-}
-
-compileKotlin {
-    kotlinOptions.jvmTarget = "1.8"
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = "1.8"
-}
-
-test {
-    useJUnitPlatform()
-}
