+++ build.gradle.kts
@@ -1,0 +1,7 @@
+buildscript {
+    repositories {
+        jcenter()
+        mavenCentral()
+        mavenLocal()
+    }
+}
