+++ buildSrc/src/main/kotlin/com/caseykulm/spellingbee/gradle/constants/Versions.kt
@@ -1,0 +1,9 @@
+package com.caseykulm.spellingbee.gradle.constants
+
+object Versions {
+    const val junit5 = "5.6.2"
+    const val kotlin = "1.4.10"
+    const val kotlinx = "1.3.9"
+    const val ktlint = "0.39.0"
+    const val moshi = "1.11.0"
+}
