+++ lib/src/main/kotlin/models/BoardSolutions.kt
@@ -1,0 +1,4 @@
+package models
+
+typealias BoardSolutions = Map<SpellingBeeBoard, Set<String>>
+typealias BoardSolution = Pair<SpellingBeeBoard, Set<String>>
