+++ lib/src/main/kotlin/models/Dictionary.kt
@@ -1,0 +1,6 @@
+package models
+
+typealias Dictionary = Map<String, String>
+
+fun Dictionary.definition(word: String): String? = this[word]
+val Dictionary.words get() = keys
