+++ lib/src/main/kotlin/models/SpellingBeeFilterRule.kt
@@ -1,0 +1,13 @@
+package models
+
+import processor.UniqueCharSetPool
+
+enum class SpellingBeeFilterRule(val predicate: (String, UniqueCharSetPool) -> Boolean) {
+    AlphaOnly(predicate = { word: String, _: UniqueCharSetPool -> word.contains(Regex("^[A-z]+\$")) }),
+    Length4OrGreater(predicate = { word: String, _ -> word.length >= 4 }),
+    SevenOrFewerUnique(
+        predicate = { word: String, ucsPool: UniqueCharSetPool ->
+            ucsPool.getOrCreateUniqueCharSet(word).uniqueCount <= 7
+        }
+    )
+}
