+++ lib/src/main/kotlin/processor/BoardSolutionsProcessor.kt
@@ -1,0 +1,8 @@
+package processor
+
+import models.BoardSolutions
+import models.Dictionary
+
+interface BoardSolutionsProcessor {
+    fun process(dictionary: Dictionary): BoardSolutions
+}
