+++ lib/src/main/resources/105_solutions.json
