--- src/main/kotlin/TimeTracking.kt
@@ -1,8 +1,0 @@
-fun <T> timing(message: String, pipe: () -> T): T {
-    val startTime = System.currentTimeMillis()
-    val any = pipe()
-    val totalTime: Long = System.currentTimeMillis() - startTime
-    val totalTimeStr = totalTime.toString().padStart(length = 5, padChar = ' ')
-    println("$totalTimeStr ms to $message")
-    return any
-}
