--- src/main/kotlin/models/BoardSolutions.kt
@@ -1,4 +1,0 @@
-package models
-
-typealias BoardSolutions = Map<SpellingBeeBoard, Set<String>>
-typealias BoardSolution = Pair<SpellingBeeBoard, Set<String>>
