--- src/main/kotlin/models/Dictionary.kt
@@ -1,6 +1,0 @@
-package models
-
-typealias Dictionary = Map<String, String>
-
-fun Dictionary.definition(word: String): String? = this[word]
-val Dictionary.words get() = keys
