--- src/main/kotlin/processor/BoardSolutionsProcessor.kt
@@ -1,8 +1,0 @@
-package processor
-
-import models.BoardSolutions
-import models.Dictionary
-
-interface BoardSolutionsProcessor {
-    fun process(dictionary: Dictionary): BoardSolutions
-}
